refactor(e2e): remove backward-compat re-export shims from llm_judge and subtest_executor
Redirect all import call sites to their canonical source modules
(build_pipeline, pipeline_scripts, llm_judge_models, parallel_executor)
and remove the re-export blocks and __getattr__ shim that were bridging
the original god-class decomposition.

Closes #1685
