+++ src/scylla/e2e/rerun_judges.py
@@ -342,7 +342,8 @@ def _rerun_single_judge_slot(
         import time
         from datetime import datetime, timezone
 
-        from scylla.e2e.llm_judge import _call_claude_judge, _parse_judge_response, _save_judge_logs
+        from scylla.e2e.llm_judge import _call_claude_judge, _parse_judge_response
+        from scylla.e2e.pipeline_scripts import _save_judge_logs
 
         try:
             judge_start = time.time()
